+++ saleor/plugins/manager.py
@@ -1169,6 +1169,16 @@ class PluginsManager(PaymentInterface):
             "gift_card_deleted", default_value, gift_card
         )
 
+    def gift_card_sent(self, gift_card: "GiftCard", channel_slug: str, email: str):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "gift_card_sent",
+            default_value,
+            gift_card,
+            channel_slug,
+            email,
+        )
+
     def gift_card_status_changed(self, gift_card: "GiftCard"):
         default_value = None
         return self.__run_method_on_plugins(